Update Terrapin Station.txt
Just a few chord corrections - well, I think corrections. My observations are based on the original studio version.

The line Inspiration . . . has a Dmaj7 after the D at the end of the line.

The C#m chords in the song after the modulation to A have been revised to C#m7 chords,

The repeated chords in the instrumental section at the end of the song have some suggested revisions.